Mixins
* Break up the style transformer into mixins so that there isn't just
  one giant class
